+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.4.0 - 2024-01-09
+
+1. Numeric property handling for feature flags now does the expected: When passed in a number, we do a numeric comparison. When passed in a string, we do a string comparison. Previously, we always did a string comparison.
+2. Add support for relative date operators for local evaluation.
+
 # 2.3.1 - 2023-08-14
 
 1. Update option doc string to show personal API Key as an option